feat: Adjustable-width canvas element
Three.js determines the aspect ratio of the canvas element differently
depending on the camera in use.

This spline scene uses the [THREE.OrthographicCamera][1], which uses
left, right, top, and bottom arguments to determine the aspect ratio
rendered inside the 3d scene rendered to the canvas element.

By default, Spline assumes the viewport of the canvas element should be
the full browser width and height, so if the canvas element is styled to
be a different aspect ratio than the intrinsic viewport with CSS, the
scene rendered inside the canvas element will appear distorted.

This demo uses [lil-gui][2] to provide a debug interface to a `width`
ratio variable, which we can adjust between 0–1 to represent the
proportional width of the browser the canvas should occupy. This
variable is referenced when the three js orthographic camera is
instantiated on initial render, and whenever the browser is resized.

This allows us to avoid the canvas being squashed by CSS and
providing a squashed looking render.

[1]: https://threejs.org/docs/#api/en/cameras/OrthographicCamera
[2]: https://github.com/georgealways/lil-gui